[ADD] add new weight property to all blocks. - Add NakamotoHeaviestChainConsensus
Former-commit-id: 81db75fe404897d1348df840733a28f080aa2109 [formerly 94476e9ee33558b86285b9e73c64aa9fae8dbf2d]
Former-commit-id: 80beb583be26abfb4b131fcb4c06f79181181b8d

+++ src/main/java/jabs/consensus/algorithm/TangleIOTA.java
@@ -12,7 +12,7 @@ import java.util.HashSet;
  *
  */
 public class TangleIOTA extends AbstractDAGBasedConsensus<TangleBlock, TangleTx> {
-    protected final HashMap<TangleBlock, Integer> blockAccWeights = new HashMap<>();
+    protected final HashMap<TangleBlock, Double> blockAccWeights = new HashMap<>();
 
     /**
      * Creates a Tangle Consensus Algorithm
